requirements.txt
# Core ML Dependencies
torch>=2.1.0
torch-geometric>=2.4.0
transformers>=4.35.0
sentence-transformers>=2.2.2
datasets>=2.14.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.1.0

# Graph Processing
networkx>=3.2
dgl>=1.1.0

# Geospatial
geopy>=2.4.0
folium>=0.14.0
geopandas>=0.14.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
redis>=5.0.0

# Monitoring & Observability
prometheus-client>=0.18.0
structlog>=23.1.0
sentry-sdk>=1.38.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Validation
cerberus>=1.3.4
marshmallow>=3.20.0

# Utilities
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.6.0
typer>=0.9.0

# Caching
diskcache>=5.6.0
joblib>=1.3.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Model Serving
onnx>=1.15.0
onnxruntime>=1.16.0

# Cloud & Deployment
boto3>=1.34.0
google-cloud-storage>=2.10.0
kubernetes>=28.1.0 